Explain the advanced example
We use message broadcast for convenience. When the program is started, or when the right button is
pressed, we trigger `display` and `sound` coroutines, which display a message on the light matrix
and play some sounds respectively, simultaneously. When the left button is pressed, we stop all other coroutines (so
that we don't play the second sound after the first one is interrupted, for example) and interrupt
any ongoing display and sound by clearing the display and playing no sound.